Factor collectSources & CompileError into compile-common
