+++ packages/compile-common/package.json
@@ -19,6 +19,8 @@
   },
   "dependencies": {
     "@truffle/contract-schema": "^3.3.2",
+    "@truffle/error": "^0.0.11",
+    "colors": "^1.4.0",
     "fs-extra": "^8.1.0",
     "mocha": "8.0.1",
     "ts-node": "8.10.2"
