+++ packages/compile-common/src/index.ts
@@ -1,3 +1,5 @@
 export { Profiler } from "./profiler";
 export * as Shims from "./shims";
+export * as Sources from "./sources";
+export * as Errors from "./errors";
 export * from "./types";
