+++ packages/compile-common/src/sources.ts
@@ -1,0 +1,70 @@
+import { Sources, CollectedSources } from "./types";
+
+/**
+ * Collects sources, targets into collections with OS-independent paths,
+ * along with a reverse mapping to the original path (for post-processing)
+ *
+ * @param originalSources - { [originalSourcePath]: contents }
+ * @param originalTargets - originalSourcePath[]
+ * @return { sources, targets, originalSourcePaths }
+ */
+export function collectSources(
+  originalSources: Sources,
+  originalTargets: string[] = []
+): CollectedSources {
+  const mappedResults = Object.entries(originalSources)
+    .map(([originalSourcePath, contents]) => ({
+      originalSourcePath,
+      contents,
+      sourcePath: getPortableSourcePath(originalSourcePath)
+    }))
+    .map(({ originalSourcePath, sourcePath, contents }) => ({
+      sources: {
+        [sourcePath]: contents
+      },
+
+      // include transformed form as target if original is a target
+      targets: originalTargets.includes(originalSourcePath) ? [sourcePath] : [],
+
+      originalSourcePaths: {
+        [sourcePath]: originalSourcePath
+      }
+    }));
+
+  const defaultAccumulator = {
+    sources: {},
+    targets: [] as string[],
+    originalSourcePaths: {}
+  };
+
+  return mappedResults.reduce(
+    (accumulator, result) => ({
+      sources: Object.assign({}, accumulator.sources, result.sources),
+      targets: [...accumulator.targets, ...result.targets],
+      originalSourcePaths: Object.assign(
+        {},
+        accumulator.originalSourcePaths,
+        result.originalSourcePaths
+      )
+    }),
+    defaultAccumulator
+  );
+}
+
+/**
+ * @param sourcePath - string
+ * @return string - operating system independent path
+ * @private
+ */
+function getPortableSourcePath(sourcePath: string): string {
+  // Turn all backslashes into forward slashes
+  var replacement = sourcePath.replace(/\\/g, "/");
+
+  // Turn G:/.../ into /G/.../ for Windows
+  if (replacement.length >= 2 && replacement[1] === ":") {
+    replacement = "/" + replacement;
+    replacement = replacement.replace(":", "");
+  }
+
+  return replacement;
+}
