+++ packages/compile-solidity/package.json
@@ -18,9 +18,7 @@
     "@truffle/compile-common": "^0.5.0",
     "@truffle/config": "^1.2.33",
     "@truffle/contract-sources": "^0.1.8",
-    "@truffle/error": "^0.0.11",
     "@truffle/expect": "^0.0.15",
-    "colors": "^1.1.2",
     "debug": "^4.1.0",
     "fs-extra": "^8.1.0",
     "ora": "^3.4.0",
