+++ packages/compile-vyper/package.json
@@ -11,6 +11,7 @@
     "test": "mocha"
   },
   "dependencies": {
+    "@truffle/compile-common": "^0.5.0",
     "@truffle/compile-solidity": "^5.1.1",
     "colors": "^1.1.2",
     "debug": "^4.3.1",
